refactor: move list_comports from utils.py to comm.py
